수정

+++ src/pet/pet.service.ts
@@ -40,57 +40,54 @@ export class PetService {
 
   //자신의 펫 리스트
   async MyPetList(userId: number) {
-    try{
+    try {
       const MyPet = await this.petRepository.find({ where: { userId } })
       if (MyPet.length === 0) {
         return { message: "펫이 존재하지않습니다." }
       }
+      
       return MyPet;
     }
-    catch(err){
+    catch (err) {
       throw new Error(err)
     }
 
   }
 
   //자신의 펫 상세정보
-  async MyPetDetail(id : number , userId : number){
-    try{
-      const MyPet = await this.petRepository.findOne({ where: { userId , id } })
+  async MyPetDetail(id: number, userId: number) {
+   
+      const MyPet = await this.petRepository.findOne({ where: { id } ,relations:["schedule"] })
       if (!MyPet) {
         return { message: "펫이 존재하지않습니다." }
       }
-      return MyPet;
-    }
-    catch(err){
-      throw new Error(err)
-    }
-
-
+      if(MyPet.userId === userId){
+        return MyPet;
+      }
   }
 
 
   //펫 등록
-  async create(filename: string, file: Buffer, name: string,  age: string, breed: string, birth: Date, gender : string , userId: number) {
+  async create(filename: string, file: Buffer, name: string, age: string, breed: string, birth: Date, gender: string, userId: number) {
     const ext = extname(filename);
     const baseName = basename(filename, ext);
     const filenames = `images/${baseName}-${Date.now()}${ext}`
-   
+
     try {
       await this.s3Client.send(new PutObjectCommand({ Bucket: "sunah", Key: filenames, Body: file }))
       await this.petRepository.save({ userId, profileImage: "https://sunah.s3.ap-northeast-2.amazonaws.com/" + filenames, name, age, breed, birth, gender })
-      return {message : "등록에 성공하였습니다"}
+      return { message: "등록에 성공하였습니다" }
     }
     catch (err) {
       throw new Error(err);
     }
   }
 
-  async createNotImage(name: string,  age: string, breed: string, birth: Date, gender : string , userId: number) {
+  async createNotImage(name: string, age: string, breed: string, birth: Date, gender: string, userId: number) {
 
     try {
       await this.petRepository.save({ userId, name, age, breed, birth, gender })
-      return {message : "등록에 성공하였습니다"}
+      return { message: "등록에 성공하였습니다" }
     }
     catch (err) {
       throw new Error(err);
@@ -98,45 +95,51 @@ export class PetService {
 
   }
 
-  
+
 
 
   //펫 정보 수정
-  async PetUpdate(id : number, userId : number, filename: string, file: Buffer, name: string,  age: string, breed: string, birth: Date, gender : string ,){
+  async PetUpdate(id: number, userId: number, filename: string, file: Buffer, name: string, age: string, breed: string, birth: Date, gender: string,) {
     const ext = extname(filename);
     const baseName = basename(filename, ext);
     const filenames = `images/${baseName}-${Date.now()}${ext}`
-    try {
+    const pet = await this.petRepository.findOne({ where: { id } });
+    if (pet.userId === userId) {
       await this.s3Client.send(new PutObjectCommand({ Bucket: "sunah", Key: filenames, Body: file }))
-      await this.petRepository.update({id, userId},{ userId, profileImage:"https://sunah.s3.ap-northeast-2.amazonaws.com/" + filenames, name, age, breed, birth, gender })
-      return {message : "수정에 성공하였습니다"}
+      await this.petRepository.update({ id, userId }, { userId, profileImage: "https://sunah.s3.ap-northeast-2.amazonaws.com/" + filenames, name, age, breed, birth, gender })
+      return { message: "수정에 성공하였습니다" }
+
     }
-    catch (err) {
-      throw new Error(err);
-    }
+    return { message: "수정 권한이 없습니다." }
+
+
 
 
   }
 
-  async PetUpdateNotimage(id : number , userId : number, name : string, age : string, breed : string, birth : Date, gender : string){
-    try {
-      await this.petRepository.update({id, userId},{ userId, name, age, breed, birth, gender })
-      return {message : "수정에 성공하였습니다"}
+  async PetUpdateNotimage(id: number, userId: number, name: string, age: string, breed: string, birth: Date, gender: string) {
+
+    const pet = await this.petRepository.findOne({ where: { id } });
+    if (pet.userId === userId) {
+      await this.petRepository.update({ id, userId }, { userId, name, age, breed, birth, gender })
+      return { message: "수정에 성공하였습니다" }
     }
-    catch (err) {
-      throw new Error(err);
-    }
+    return { message: "수정 권한이 없습니다." }
 
   }
 
   //펫 삭제
-  async PetDelete(id : number , userId : number){
-    const Pet = await this.petRepository.findOne({where : {id , userId}});
-    if(!Pet){
+  async PetDelete(id: number, userId: number) {
+    const Pet = await this.petRepository.findOne({ where: { id } });
+    if (!Pet) {
       throw new HttpException('펫을 찾을 수 없습니다.', HttpStatus.NOT_FOUND);
     }
-    await this.scheduleRepository.delete({petId : id, userId})
-    await this.petRepository.delete({id , userId});
-    return {message : "삭제에 성공했습니다"}
+    if (Pet.id === userId) {
+      await this.scheduleRepository.delete({ petId: id, userId })
+      await this.petRepository.delete({ id, userId });
+      return { message: "삭제에 성공했습니다" }
+    }
+    return { message: "삭제 권한이 없습니다" }
+
   }
 }
